# docs/technical/openapi-v3.yaml
openapi: 3.1.0
info:
  title: WebWMS API v3
  version: 3.0.0
  description: Tenant-scoped API for master data, warehouse processes and integrations.
servers:
  - url: /api/v3
security:
  - ApiKey: []
paths:
...
  /integration/import:
    post:
      summary: Import JSON, XML, CSV, XLSX or SAP IDoc
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportRequest'
      responses:
        '201': { $ref: '#/components/responses/Item' }
        '422': { $ref: '#/components/responses/Problem' }
  /integration/export:
    post:
      summary: Export records as JSON, XML, CSV or XLSX
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '201': { $ref: '#/components/responses/Item' }
...
  /integration/commerce-orders:
    post:
      summary: Import a channel order idempotently by connection and external order ID
      responses:
        '201': { $ref: '#/components/responses/Item' }
        '409': { $ref: '#/components/responses/Problem' }
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    ImportRequest:
      type: object
      required: [format, resourceType, content]
      properties:
        format: { type: string, enum: [json, xml, csv, xlsx, idoc] }
        resourceType: { type: string }
        sourceReference: { type: [string, 'null'] }
        content: { type: string, description: XLSX content is base64 encoded. }
    ExportRequest:
      type: object
      required: [format, resourceType, rows]
      properties:
        format: { type: string, enum: [json, xml, csv, xlsx] }
        resourceType: { type: string }
        rows:
          type: array
          items: { type: object, additionalProperties: true }
    Problem:
      type: object
      required: [type, title, status, detail]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
  responses:
...
    Item:
      description: Tenant-scoped resource
      content:
        application/json:
          schema:
            type: object
            required: [data]
            properties:
              data: { type: object, additionalProperties: true }
    Problem:
      description: RFC-7807-style error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }